refactor ntl outage
