+++ rapida/ntl/outage.py
@@ -3,34 +3,17 @@ from datetime import datetime, timedelta
 import numbers
 import logging
 from rich.progress import Progress
-from rapida.ntl.util import write_outage_tif
+from rapida.ntl.utils import write_outage_tif
 from rapida.ntl.fetch import fetch
 from rapida.ntl.nasa.io import  extract_bb
 from rapida.ntl.nasa import const as nasa_const
 import numpy as np
 import os
-from scipy.ndimage import label
-from rapida.ntl.util import calculate_regional_outage_simplified
+from rapida.ntl.utils import calculate_regional_outage_simplified
 
 
 
 logger = logging.getLogger('rapida')
-
-
-
-def spatial_filter(outage_map, min_size=2):
-    # 1. Group connected pixels into "clumps"
-    labeled_array, num_features = label(outage_map)
-
-    # 2. Count how many pixels are in each clump
-    clump_sizes = np.bincount(labeled_array.ravel())
-
-    # 3. Create a mask of clumps that meet your size requirement
-    mask_size = clump_sizes >= min_size
-
-    # 4. Filter the original map (clump 0 is the background, so we ignore it)
-    mask_size[0] = 0
-    return mask_size[labeled_array]
